Fix incorrect pin setup and usage.
Prior to this change the script setup, and used, pin 11 to turn the buzzer on and off.

The circuit when constructed according to the documentation should utilise pin 12.

Additionally, the script is altered to ensure that the pin is setup off, so the buzzer is off, then the pin is switched on and off in the main loop.