Add setup() and loop() functions
The setup() function is added and script initialisation code refactored
to setup().

The loop() function is also added that implements the "main" loop for this
script. The logic that changes the frequency of the buzzer at defined
intervals has been moved into the loop() function.

Finally, the script is restructured such that setup() and loop() functions
are only called by the script if it is the main program being executed
by the python interpreter.