Change GPIO pins used for tilt swith and LED.
The circuit built from the Adeept kit is constructed to use GPIO pin 11
for the LED and GPIO pin 12 for the tilt switch. Before the change the
script was written to use GPIO pin 12 for the LED an GPIO pin 11 for the
tilt switch. This change corrects the pin setup and usage by ensuring that
GPIO pin 11 is used for the LED and GPIO pin 12 is used for the tilt switch.