Correct pin numbering and usage for switch that controls relay.
The circuit for lesson 06 in the Adeept Learning kit utilises a
switch that controls the relay. Prior to the change the script
designated, via the RelayPin variable, the RPi GPIO pin that controls
the switch as pin 11. The correct pin controlling the switch,
consistent with documentation, is pin 12. This change corrects
the pin used to control the switch controlling the relay. Additionally,
the name of the ReplayPin variable is changed to switch_pin to reflect
it controls the switch that controls current supplied to the relay.

In addition to correcting the pin numbering this change corrects
the logic in the loop() used to control the switch. The switch is
turned on by making GPIO pin 12 go HIGH, and turned off by going LOW.