Correct incorrect pin numbering and correct implementation.
The pin numbers specified in the code were not consistent with
the circuit diagram presented in the kit for the lesson. This change
corrects the usage. Additionally, the code is altered in order to
allow the caller to obtain changes for both the x-axis and y-axis of
the joystick, rather than only a single channel/axis as previously
implemented.