Correct naming of destroy functioning and improve efficiency.
The implementation of getResult() is changed to improve efficiency.
The previous implementation wasted half of the reads of the joystick,
by reading the x-axis then returning if it was left, and if it wasn't
reading the x-axis again. The implementation now reads the x-axis
and return either left or right. If joystick wasn't moved in the left
or right direction, the y-axis is read in a similar manner returning
up/down. Additionally, a short sleep period is added, to slow down
reading joystick movement to prevent cpu saturation in polling for
joystick movement.

Finally, button press detection is changed to using interrupt detection
to improve efficicency.