Resolve authentication error.
Prior to the change swiping an RFID card/tag produced an authentication error.
This change appears to resolve the error. The issue is an integration
problem of the modules, in particular the buzzer module. Moving the calls
for the buzzer until after the RFID module authenticates the card appears
to have resolved the issue.